Fix Order

+++ app/Http/Controllers/OrderController.php
@@ -295,10 +295,12 @@ class OrderController extends Controller
         // Se l'ordine è stato aggiornato, elimina gli OrderProduct associati
         if ($request->id) {
             $orderProducts = OrderProduct::where('id_order', $request->id);
-            foreach ($orderProducts as $product) {
-                ProductVariant::where('id', $product->id_product_variant)->increment('stock', $product->quantity);
+            if($orderProducts) {
+                foreach ($orderProducts as $product) {
+                    ProductVariant::where('id', $product->id_product_variant)->increment('stock', $product->quantity);
+                }
+                $orderProducts->delete();
             }
-            $orderProducts->delete();
         }
 
         // Itera attraverso i prodotti dell'ordine
